test2

+++ ft_memcpy.c
@@ -6,7 +6,7 @@
 /*   By: jaeyjeon <@student.42seoul.kr>             +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/20 16:10:57 by jaeyjeon          #+#    #+#             */
-/*   Updated: 2021/11/20 17:24:36 by jaeyjeon         ###   ########.fr       */
+/*   Updated: 2021/11/20 19:43:34 by jaeyjeon         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void    *ft_memcpy(void *dst, const void *src, size_t n)
 {
-    int i;
+    size_t i;
 
     i = 0;
     if (dst == NULL && src == NULL)
